nits: Move loss to separate file + address review feedback (Gemini, Sachin)
